requirements.txt
environs>=9.5.0,<10.0.0
Flask>=2.1.2
attrs>=20.3.0,<21.0.0
retrying>=1.3.3,<2.0.0
aiohttp>=3.8.1,<4.0.0
requests>=2.25.1,<3.0.0
loguru>=0.5.3,<1.0.0
pyquery>=1.4.3,<2.0.0
supervisor>=4.2.1,<5.0.0
redis>=3.5.3,<4.0.0
lxml>=4.6.5,<5.0.0
fake_headers>=1.0.2,<2.0.0
maxminddb_geolite2==2018.703
gevent>=21.1.0,<22.0.0
tornado>=6.0,<7.0
meinheld>=1.0.0,<2.0.0
itsdangerous>=0.24,<1.0.0
MarkupSafe<2.1.0
